authenticate

+++ app/Http/Controllers/TasksController.php
@@ -52,6 +52,9 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
+        if($task->user_id !== \Auth::user()->id) {
+            return redirect('/');
+        }
         
         return view('tasks.show', [
             'task' => $task,
@@ -62,6 +65,10 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::find($id);
+        
+        if($task->user_id !== \Auth::user()->id) {
+            return redirect('/');
+        }
         
         return view('tasks.edit', [
             'task' => $task,
@@ -76,6 +83,10 @@ class TasksController extends Controller
         ]);
         
         $task = Task::find($id);
+        if($task->user_id !== \Auth::user()->id) {
+            return redirect('/');
+        }
+        
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
@@ -88,6 +99,9 @@ class TasksController extends Controller
     {
         
         $task = Task::find($id);
+        if($task->user_id !== \Auth::user()->id) {
+            return redirect('/');
+        }
         $task->delete();
         
         return redirect('/');
